renamed iceberg plugin -> sync plugin
