--- plugins/iceberg/README.md
@@ -1,3 +1,0 @@
-# Howler Iceberg Plugin
-
-This plugin contains modules for Iceberg data sync from Howler.
