+++ plugins/sync/.coveragerc.pytest
