+++ plugins/sync/.gitignore
