+++ plugins/sync/LICENSE
