+++ plugins/sync/README.md
@@ -1,0 +1,3 @@
+# Howler Sync Plugin
+
+This plugin contains modules for data sync from Howler.
