+++ plugins/sync/build_scripts/set_version.py
