+++ plugins/sync/setup.cfg
