+++ plugins/sync/test/unit/sync_service/test_get_upserted_hits.py
@@ -2,14 +2,14 @@ from datetime import timedelta
 
 
 def test_no_interval_gets_all_hits():
-    from iceberg.services import sync_service
+    from sync.services import sync_service
 
     hits = sync_service.get_upserted_hits()
     assert len(hits) == 20
 
 
 def test_interval_start_get_hits(current_time):
-    from iceberg.services import sync_service
+    from sync.services import sync_service
 
     start_time = current_time - timedelta(days=1)
     hits = sync_service.get_upserted_hits(data_interval_start=start_time)
@@ -17,7 +17,7 @@ def test_interval_start_get_hits(current_time):
 
 
 def test_interval_start_with_updates(current_time):
-    from iceberg.services import sync_service
+    from sync.services import sync_service
 
     start_time = current_time - timedelta(days=2)
     hits = sync_service.get_upserted_hits(data_interval_start=start_time)
@@ -25,7 +25,7 @@ def test_interval_start_with_updates(current_time):
 
 
 def test_interval_start_and_end_get_hits(current_time):
-    from iceberg.services import sync_service
+    from sync.services import sync_service
 
     start_time = current_time - timedelta(days=3)
     end_time = current_time - timedelta(days=2, hours=1)
@@ -34,7 +34,7 @@ def test_interval_start_and_end_get_hits(current_time):
 
 
 def test_interval_start_and_end_lower_bound_is_inclusive(current_time):
-    from iceberg.services import sync_service
+    from sync.services import sync_service
 
     start_time = current_time - timedelta(days=1)
     end_time = current_time
